requirements.txt
asgiref==3.11.0
coverage==7.13.1
Django==6.0.1
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
